chore(core/modules): move process manager to process sub directory
