+++ apps/core/src/types/routing.ts
@@ -1,6 +1,6 @@
 import type { FastifyPluginAsync, FastifyRequest } from 'fastify';
 import 'fastify';
-import type { ProcessManager } from '../modules/process.manager';
+import type { ProcessManager } from '../modules/process/manager';
 import type { GameManager } from '../modules/game/manager';
 
 export interface Managers {
